refactored back button navigation
