+++ src/pages/AddRoom.jsx
@@ -3,9 +3,9 @@ import {useCreateNewVespaMutation} from '../redux/vespaAPI';
 
 import { Toast, useToast } from '../components/Toast';
 
-import {useNavigate} from 'react-router-dom';
 
 import backimg from '../assets/background.jpg';
+import BackButton from '../components/BackButton';
 
 
 
@@ -17,7 +17,6 @@ function AddRoom() {
   const [createVespa, {isLoading: isCreating}] = useCreateNewVespaMutation();
 
   const [displayBool, message, type, showToast] = useToast();
-  const navigate = useNavigate();
 
   let initialVespaData = {
     Name: '',
@@ -37,9 +36,7 @@ function AddRoom() {
 
   }
 
-  const handleGoBack = () => {
-    navigate('/home');
-  };
+
   const handleSubmit = (e) => {
     e.preventDefault()
 
@@ -68,10 +65,9 @@ function AddRoom() {
 
       {displayBool && <Toast message={message} type={type}/>}
 
-      <button className='z-10 absolute left-0 top-0 m-4  ' onClick={handleGoBack}>
 
-<svg className="h-20 w-20 text-black"  viewBox="0 0 24 24"  fill="none"  stroke="currentColor"  strokeWidth="2"  strokeLinecap="round"  strokeLinejoin="round">  <circle cx="12" cy="12" r="10" />  <polyline points="12 8 8 12 12 16" />  <line x1="16" y1="12" x2="8" y2="12" /></svg>
-</button>
+      <BackButton />
+
       <div className="absolute inset-0 overflow-hidden">
         <img src={backimg} alt="Background" className="h-full w-full object-fill " />
         <div className="absolute inset-0 z-0 opacity-90 bg-[#96bf01]" />
